frontend: badges: move badge code to badges.py
